Update Config class for access by dot notation
Major refactor allows for access to nested configuration items by dot
notation. For example, a config of the format `{'a':{'b':{'c':1}}}`
could now be accessed (in Config() instance A) via A.a.b.c.

Given that the config system is the main point of contact between the
training loop and the user, it is important that the user be able to
straightforwardly define access to config variables in the training
loop. In other words, no messing with nested dictionaries. All config
options can be accessed through dot notation.

Lastly, config is maintained as a dictionary called 'config.' This can
always be accessed with A.config.