Add config printing and config type checking to Config class
In an effort to encourage future extensibility, I expanded the
configuration checker to check if the input is a dict or yaml and then
to return a TypeError noting that one of these two inputs is
needed. It will be fairly trivial to add, e.g., JSON support in the
future.

I also added a method to print and optionally save the config file to
a specified directory.

I deleted the deprecated code.

With this, the config system has all of the functionality we need.