remove old notebooks
This commit removes old notebooks that do not demonstrate core elements of the functionality offered
in this package. This is the first step in a broader package overhaul. It was initially conceived of
as an all-inclusive way to organize and run numerai models. I have re-envisioned it as a simpler
repo of utilities for handling some of the core functionality of (especially) data handling and
cross validation.